More layout improvements + prettier

+++ src/main/app/src/views/MittausForm.tsx
@@ -1,112 +1,129 @@
-import React from 'react'
-import { Formik, FieldArray as FFieldArray } from 'formik'
+import React from "react";
+import { Formik, FieldArray as FFieldArray } from "formik";
 import * as Yup from "yup";
-import MittausStore from '../stores/MittausStore'
-import { FormikCustomDatePicker } from '../components/CustomDatePicker';
-import { CustomNumber } from '../components/CustomNumber';
-import { CustomText } from '../components/CustomText';
-import AsennettuAnturiForm from './AsennettuAnturiForm';
-import AsennettuAnturiStore from '../stores/AsennettuAnturiStore';
-import { Form as FForm } from 'formik';
-import { Button } from 'react-bootstrap';
-import SeliteTypeEnum from '../types/enums/seliteType.enum';
-import MittausSuuntaTypeEnum from '../types/enums/mittausSuuntaType.enum';
-import { useParams, useHistory } from 'react-router-dom';
-import IMittaus from '../types/interfaces/mittaus.interface';
-import { getData, deleteData, putData, postData } from '../api';
+import MittausStore from "../stores/MittausStore";
+import { FormikCustomDatePicker } from "../components/CustomDatePicker";
+import { CustomNumber } from "../components/CustomNumber";
+import { CustomText } from "../components/CustomText";
+import AsennettuAnturiForm from "./AsennettuAnturiForm";
+import AsennettuAnturiStore from "../stores/AsennettuAnturiStore";
+import { Form as FForm } from "formik";
+import { Button } from "react-bootstrap";
+import SeliteTypeEnum from "../types/enums/seliteType.enum";
+import MittausSuuntaTypeEnum from "../types/enums/mittausSuuntaType.enum";
+import { useParams, useHistory } from "react-router-dom";
+import IMittaus from "../types/interfaces/mittaus.interface";
+import { getData, deleteData, putData, postData } from "../api";
 
 const validationSchemaTunnusarvot = Yup.object({
-  mittaussuunta_xyz: Yup.mixed<string>().oneOf(Object.values(MittausSuuntaTypeEnum)).required(),
-  tarinan_maksimiarvo: Yup.number().min(0,
+  mittaussuunta_xyz: Yup.mixed<string>()
+    .oneOf(Object.values(MittausSuuntaTypeEnum))
+    .required(),
+  tarinan_maksimiarvo: Yup.number().min(0, "Arvon tulee olla positiivinen"),
+  hallitseva_taajuus: Yup.number().min(0, "Arvon tulee olla positiivinen"),
+  tarinan_tunnusluku_vw95_rms: Yup.number().min(
+    0,
     "Arvon tulee olla positiivinen"
   ),
-  hallitseva_taajuus: Yup.number().min(0,
-    "Arvon tulee olla positiivinen"
-  ),
-  tarinan_tunnusluku_vw95_rms: Yup.number().min(0,
-    "Arvon tulee olla positiivinen"
-  )
-})
-
+});
 
 const validationSchemaAsennuspaikanTyyppi = Yup.object({
   selite: Yup.mixed<string>().oneOf(Object.values(SeliteTypeEnum)).required(),
-  lisatiedot: Yup.string().when(['selite'], {
-    is: ( selite: string ) => selite === SeliteTypeEnum.muu,
-    then: Yup.string().required('Lisätiedot vaaditaan, kun sijoituspaikka on "muu"'),
-    otherwise: Yup.string().nullable()
-  })
-})
+  lisatiedot: Yup.string().when(["selite"], {
+    is: (selite: string) => selite === SeliteTypeEnum.muu,
+    then: Yup.string().required(
+      'Lisätiedot vaaditaan, kun sijoituspaikka on "muu"'
+    ),
+    otherwise: Yup.string().nullable(),
+  }),
+});
 
 const validationSchemaAsennettuAnturi = Yup.object({
   gps_lat: Yup.number()
-    .typeError('Koordinaatin tulee olla numeerinen')
-    .required('Koordinaatit eivät voi olla tyhjiä'),
+    .typeError("Koordinaatin tulee olla numeerinen")
+    .required("Koordinaatit eivät voi olla tyhjiä"),
   gps_long: Yup.number()
-    .typeError('Koordinaatin tulee olla numeerinen')    
-    .required('Koordinaatit eivät voi olla tyhjiä'),
-  etaisyys_radasta_jos_eri: Yup.number().min(0, "Etäisyyden tulee olla positiivinen")
+    .typeError("Koordinaatin tulee olla numeerinen")
+    .required("Koordinaatit eivät voi olla tyhjiä"),
+  etaisyys_radasta_jos_eri: Yup.number()
+    .min(0, "Etäisyyden tulee olla positiivinen")
     .typeError("Etäisyyden tulee olla numero")
     .required("Etäisyys vaaditaan"),
-  kerros: Yup.number().integer()
+  kerros: Yup.number()
+    .integer()
     .typeError("Kerroksen tulee olla numero")
     .required("Kerros vaaditaan"),
   sijoituspaikan_lisaselite: Yup.string().nullable(),
   asennuspaikanTyyppi: validationSchemaAsennuspaikanTyyppi,
-  anturikohtaisetTunnusarvot: Yup.array().of(validationSchemaTunnusarvot).max(3)
-})
+  anturikohtaisetTunnusarvot: Yup.array()
+    .of(validationSchemaTunnusarvot)
+    .max(3),
+});
 
-const validationSchema = Yup.object().shape({
-  alkuaika: Yup.date().typeError("Alkuajan tulee olla päivämäärä")
-    .required("Alkuaika vaaditaan"),
-  loppuaika: Yup
-    .date()
-    .min(
-      Yup.ref('alkuaika'),
-      "Loppuaika ei voi olla ennen alkuaikaa"
-    )
-    .typeError("Loppuajan tulee olla päivämäärä")
-    .required("Loppuaika vaaditaan"),
-  mittaus_asianhallinta_id: Yup.string().trim()
-    .when('pdf_raportin_linkki', {
-      is: (val: any) => !!val,
-      then: Yup.string(),
-      otherwise: Yup.string().required('Joko asianhallintatunnus tai pdf-raportin linkki vaaditaan')
+const validationSchema = Yup.object().shape(
+  {
+    alkuaika: Yup.date()
+      .typeError("Alkuajan tulee olla päivämäärä")
+      .required("Alkuaika vaaditaan"),
+    loppuaika: Yup.date()
+      .min(Yup.ref("alkuaika"), "Loppuaika ei voi olla ennen alkuaikaa")
+      .typeError("Loppuajan tulee olla päivämäärä")
+      .required("Loppuaika vaaditaan"),
+    mittaus_asianhallinta_id: Yup.string()
+      .trim()
+      .when("pdf_raportin_linkki", {
+        is: (val: any) => !!val,
+        then: Yup.string(),
+        otherwise: Yup.string().required(
+          "Joko asianhallintatunnus tai pdf-raportin linkki vaaditaan"
+        ),
+      }),
+    pdf_raportin_linkki: Yup.string()
+      .trim()
+      .when("mittaus_asianhallinta_id", {
+        is: (val: any) => !!val,
+        then: Yup.string(),
+        otherwise: Yup.string().required(
+          "Joko asianhallintatunnus tai pdf-raportin linkki vaaditaan"
+        ),
+      }),
+    rakennuksen_pinta_ala: Yup.number()
+      .typeError("Pinta-alan tulee olla numero")
+      .positive("Pinta-alan tulee olla positiivinen"),
+    julkisivumateriaali: Yup.string().trim(),
+    runkomateriaali: Yup.string().trim(),
+    perustamistapa: Yup.string().trim(),
+    rakennusvuosi: Yup.number()
+      .positive()
+      .integer()
+      .min(1500, "Liian pieni vuosiluku")
+      .max(new Date().getFullYear(), "Liian suuri vuosiluku"),
+    katuosoite: Yup.string().trim(),
+    postinumero: Yup.string()
+      .trim()
+      .matches(/^[0-9]+$/, "Postinumero ei voi sisältää kirjaimia")
+      .min(5, "Täytyy olla 5 numeroa")
+      .max(5, "Täytyy olla 5 numeroa"),
+    created_by_lx: Yup.string().trim(),
+    asennettuAnturi: Yup.array().of(validationSchemaAsennettuAnturi),
+  },
+  [["mittaus_asianhallinta_id", "pdf_raportin_linkki"]]
+);
+
+const initializeEmptyFields = (mittaus: IMittaus): IMittaus =>
+  Object.entries(mittaus).reduce(
+    (acc, [k, v]) => ({
+      ...acc,
+      [k]: v || "",
     }),
-  pdf_raportin_linkki: Yup.string().trim()
-  .when('mittaus_asianhallinta_id', {
-    is: (val: any) => !!val,
-    then: Yup.string(),
-    otherwise: Yup.string().required('Joko asianhallintatunnus tai pdf-raportin linkki vaaditaan')
-  }),
-  rakennuksen_pinta_ala: Yup.number()
-    .typeError("Pinta-alan tulee olla numero")
-    .positive("Pinta-alan tulee olla positiivinen"),
-  julkisivumateriaali: Yup.string().trim(),
-  runkomateriaali: Yup.string().trim(),
-  perustamistapa: Yup.string().trim(),
-  rakennusvuosi: Yup.number().positive().integer()
-    .min(1500, "Liian pieni vuosiluku")
-    .max(new Date().getFullYear(), "Liian suuri vuosiluku"),
-  katuosoite: Yup.string().trim(),
-  postinumero: Yup.string().trim()
-    .matches(/^[0-9]+$/, "Postinumero ei voi sisältää kirjaimia")
-    .min(5, 'Täytyy olla 5 numeroa')
-    .max(5, 'Täytyy olla 5 numeroa'),
-  created_by_lx: Yup.string().trim(),
-  asennettuAnturi: Yup.array().of(validationSchemaAsennettuAnturi)
-}, [['mittaus_asianhallinta_id', 'pdf_raportin_linkki']])
-
-
-const initializeEmptyFields = (mittaus: IMittaus): IMittaus => Object.entries(mittaus)
-  .reduce((acc, [k, v]) => ({
-  ...acc,
-  [k]: v || "",
-}), {} as IMittaus);
+    {} as IMittaus
+  );
 
 const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
-  const [fetchedValues, setFetchedValues] = React.useState<IMittaus | null>(null);
-  console.log("mittausform", mittaus)
+  const [fetchedValues, setFetchedValues] = React.useState<IMittaus | null>(
+    null
+  );
+  console.log("mittausform", mittaus);
   const { id } = useParams<{ id: string }>();
 
   React.useEffect(() => {
@@ -114,10 +131,10 @@ const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
       const data = await getData(id);
       if (!data) {
         setFetchedValues(null);
-        return; 
+        return;
       }
       setFetchedValues(initializeEmptyFields(data));
-    }
+    };
     id && fetchAndSetData();
   }, [id]);
 
@@ -126,53 +143,55 @@ const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
   const onClickDelete = (id: string) => {
     // TODO: replace with material ui modal
     // eslint-disable-next-line no-restricted-globals
-    if (confirm('Haluatko varmasti poistaa mittauksen?'))
-      deleteData(id)
-      history.push('/mittauslista');
-  }
-  
+    if (confirm("Haluatko varmasti poistaa mittauksen?")) deleteData(id);
+    history.push("/mittauslista");
+  };
+
   const onClickUpdate = (values: any) => {
-    putData(values)
+    putData(values);
     history.go(0);
-  }
-  
+  };
 
   if (id && !fetchedValues) return <div>ei mittausta</div>;
   return (
     <>
-      <h2 style={{ marginBottom: '30px' }}>Mittauksen tiedot: {id ? `kohdetunnus ${id}` : 'Uusi mittaus'}</h2>
+      <h2 style={{ marginBottom: "30px" }}>
+        Mittauksen tiedot: {id ? `kohdetunnus ${id}` : "Uusi mittaus"}
+      </h2>
       <Formik
-        initialValues={ fetchedValues || {
-          alkuaika: '',
-          loppuaika: '',
-          mittaus_asianhallinta_id: '',
-          pdf_raportin_linkki: '',
-          rakennuksen_pinta_ala: '',
-          perustamistapa: '',
-          julkisivumateriaali: '',
-          runkomateriaali: '',
-          rakennusvuosi: '',
-          katuosoite: '',
-          postinumero: '',
-          created_by_lx: '',
-          asennettuAnturi: [],
-        }}
+        initialValues={
+          fetchedValues || {
+            alkuaika: "",
+            loppuaika: "",
+            mittaus_asianhallinta_id: "",
+            pdf_raportin_linkki: "",
+            rakennuksen_pinta_ala: "",
+            perustamistapa: "",
+            julkisivumateriaali: "",
+            runkomateriaali: "",
+            rakennusvuosi: "",
+            katuosoite: "",
+            postinumero: "",
+            created_by_lx: "",
+            asennettuAnturi: [],
+          }
+        }
         validationSchema={validationSchema}
         enableReinitialize
         validateOnMount
         onSubmit={(values, { setSubmitting }) => {
           postData({ ...values })
-          .then(res => {
-            setSubmitting(false);
-            history.push('/mittauslista');
-          })
-          .catch(err => {
-            console.log(err);
-            setSubmitting(false);
-          })
+            .then((res) => {
+              setSubmitting(false);
+              history.push("/mittauslista");
+            })
+            .catch((err) => {
+              console.log(err);
+              setSubmitting(false);
+            });
         }}
       >
-        {formik => (
+        {(formik) => (
           <FForm onSubmit={formik.handleSubmit}>
             <FormikCustomDatePicker
               label="Mittauksen alkuaika"
@@ -194,33 +213,44 @@ const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
               name="pdf_raportin_linkki"
               readOnly={false}
             />
-            {/* tähän anturi komponentit */}
+            {/* tähän anturikomponentit */}
             <FFieldArray
               name="asennettuAnturi"
-              render={arrayHelpers => (
+              render={(arrayHelpers) => (
                 <div>
-                  {formik.values.asennettuAnturi && formik.values.asennettuAnturi.length > 0 && (
+                  {formik.values.asennettuAnturi &&
+                    formik.values.asennettuAnturi.length > 0 &&
                     formik.values.asennettuAnturi.map((anturi, index) => (
                       <div key={index}>
-                        <AsennettuAnturiForm asennettuAnturi={anturi} namespace={`asennettuAnturi.${index}`} index={index} />
+                        <AsennettuAnturiForm
+                          asennettuAnturi={anturi}
+                          namespace={`asennettuAnturi.${index}`}
+                          index={index}
+                        />
                         <Button
-                          style={{ margin: '6px' }}
+                          style={{ margin: "6px" }}
                           variant="danger"
                           onClick={() => arrayHelpers.remove(index)}
                         >
                           Poista anturi
                         </Button>
                       </div>
-                    ))
-                  )}
-                  <Button style={{ margin: '6px' }} onClick={() => arrayHelpers.push(new AsennettuAnturiStore())}>
+                    ))}
+                  <Button
+                    style={{ margin: "6px" }}
+                    onClick={() =>
+                      arrayHelpers.push(new AsennettuAnturiStore())
+                    }
+                  >
                     Lisää anturi
                   </Button>
                 </div>
               )}
             />
             <h3>Mittauskohteen tiedot</h3>
-            <p>Mittaus voidaan tallentaa tietokantaan myös ilman rakennustietoja</p>
+            <p>
+              Mittaus voidaan tallentaa tietokantaan myös ilman rakennustietoja
+            </p>
             <CustomText
               label="Rakennuksen pinta-ala"
               name="rakennuksen_pinta_ala"
@@ -246,41 +276,41 @@ const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
               name="rakennusvuosi"
               readOnly={false}
             />
-            <CustomText
-              label="Katuosoite"
-              name="katuosoite"
-              readOnly={false}
-            />
+            <CustomText label="Katuosoite" name="katuosoite" readOnly={false} />
             <CustomText
               label="Postinumero"
               name="postinumero"
               readOnly={false}
             />
-            <div style={{ marginTop: '25px' }}>
+            <div style={{ marginTop: "25px" }}>
               <CustomText
                 label="Tiedot päivittänyt"
                 name="created_by_lx"
                 readOnly={true}
-                />
-              </div>
+              />
+            </div>
             <div id="button_container">
-              {id ? <>
+              {id ? (
+                <>
                   <Button
-                      type="button"
-                      variant="danger"
-                      onClick={() => {onClickDelete(id)}}
-                    >
-                      Poista
+                    type="button"
+                    variant="danger"
+                    onClick={() => {
+                      onClickDelete(id);
+                    }}
+                  >
+                    Poista
                   </Button>
                   <Button
-                      type="button"
-                      variant="primary"
-                      disabled={!formik.isValid}
-                      onClick={() => onClickUpdate(formik.values)}
-                    >
-                      Päivitä
+                    type="button"
+                    variant="primary"
+                    disabled={!formik.isValid}
+                    onClick={() => onClickUpdate(formik.values)}
+                  >
+                    Päivitä
                   </Button>
-                </> :
+                </>
+              ) : (
                 <>
                   <Button
                     type="button"
@@ -290,35 +320,35 @@ const MittausForm = ({ mittaus }: { mittaus: MittausStore }) => {
                   >
                     Tyhjennä
                   </Button>
-                  <Button type="submit" disabled={formik.isSubmitting || !formik.isValid}>
+                  <Button
+                    type="submit"
+                    disabled={formik.isSubmitting || !formik.isValid}
+                  >
                     Lähetä
                   </Button>
                 </>
-              }
+              )}
             </div>
-
           </FForm>
-
         )}
       </Formik>
     </>
-  )
-}
+  );
+};
 
-export default MittausForm
+export default MittausForm;
 
-
-export const DisplayFormikState = (props: any) =>
-  <div style={{ margin: '1rem 0' }}>
-    <h3 style={{ fontFamily: 'monospace' }} />
+export const DisplayFormikState = (props: any) => (
+  <div style={{ margin: "1rem 0" }}>
+    <h3 style={{ fontFamily: "monospace" }} />
     <pre
       style={{
-        background: '#f6f8fa',
-        fontSize: '1.25rem',
-        padding: '.5rem',
+        background: "#f6f8fa",
+        fontSize: "1.25rem",
+        padding: ".5rem",
       }}
     >
-      <strong>props</strong> ={' '}
-      {JSON.stringify(props, null, 2)}
+      <strong>props</strong> = {JSON.stringify(props, null, 2)}
     </pre>
-  </div>;
+  </div>
+);
